메인 Premium Item 이름변경, 상품목록 2열 가격정보 Ui 수정

+++ app/src/main/java/io/temco/guhada/data/model/community/CommunityDetail.kt
@@ -37,7 +37,7 @@ class CommunityDetail : Serializable {
 
     fun setBbsResponse() : CreateBbsResponse {
         var data = CreateBbsResponse()
-        data.categoryFilterId = categoryFilterId.toLong()
+        data.categoryFilterId = categoryFilterId
         data.dealName = dealName ?: ""
         data.contents = contents ?: ""
         data.brandName = brandName ?: ""
